requirements.txt
numpy==1.15.1
scikit_image==0.14.0
pandas==0.23.4
opencv-python==3.4.3.18
pillow==5.2.0
scikit-image==0.14.0
scikit_learn==0.19.2
scipy==1.1.0
tensorboardX==1.4
torch==0.4.1
torchsummary==1.4
torchvision==0.2.1
tqdm==4.25.0
